Restructure code for ToricVarieties and support more properties
